requirements.txt
attrs==25.3.0
aws-cdk.asset-awscli-v1==2.2.240
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==44.9.0
boto3==1.39.7
botocore==1.39.7
cattrs==24.1.3
constructs==10.4.2
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.112.0
publication==0.0.3
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.14.1
urllib3==2.5.0
